Updated theme template folder structure and refactored code.
