+++ src/index.js
@@ -94,8 +94,8 @@ inquirer.prompt(questionnaire).then(async (answers) => {
     //<---------------------------- For Themes integration ---------------------------------->
     if (isDark && frontEndChoice==='react') {
       fs.copyFile(
-        `${currentPath}/themeTemplates/themes.js`,
-        `${frontEnd.path}/src/themes.js`,
+        `${currentPath}/themeTemplates/react-themes/dark-theme.js`,
+        `${frontEnd.path}/src/dark-theme.js`,
         (err) => {
           if (err) {
             console.log("Error Found:", err);
@@ -105,8 +105,8 @@ inquirer.prompt(questionnaire).then(async (answers) => {
     }
     if (isDark && frontEndChoice==='vue') {
       fs.copyFile(
-        `${currentPath}/themeTemplates/Theme.vue`,
-        `${frontEnd.path}/src/Theme.vue`,
+        `${currentPath}/themeTemplates/vue-themes/dark-theme.vue`,
+        `${frontEnd.path}/src/dark-theme.vue`,
         (err) => {
           if (err) {
             console.log("Error Found:", err);
