+++ src/themeTemplates/react-themes/dark-theme.js
