+++ src/themeTemplates/vue-themes/dark-theme.vue
